Removed ability to resize the window Added function to allow Sudoku Board to grab the correct related indices based off of one index Miscellaneous modifications. Lots of things need reworking

+++ Include/Tile.hpp
@@ -14,7 +14,7 @@ public:
 	bool CollapseTile(int value);
 	void Reset();
 
-	int GetEntropy() const { return mAvailableValues.size(); }
+	size_t GetEntropy() const { return mAvailableValues.size(); }
 
 	bool operator<(const Tile& a);
